tests: Move live migration helpers to common::utils
Move the helper methods used for live migration to the common utils
(like many other tests use).
